+++ packages/infa/src/index.ts
@@ -1,9 +1,0 @@
-export type {
-  SandboxKind,
-  SandboxOptions,
-  SandboxExecOptions,
-  SandboxResult,
-  Sandbox,
-} from "./types.js";
-
-export { createSandbox } from "./factory.js";